+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # For next release
   * **Raphael Pour**
     * rework service to go with file-based db
+    * add example sentence
 
 *Not released yet*